🧪 Test Function
