+++ functions/collect_image/function.json
