+++ functions/requirements.txt
@@ -3,6 +3,7 @@ azure-core==1.25.0
 azure-functions==1.11.2
 azure-identity==1.10.0
 azure-storage-blob==12.13.1
+iot-storage-client==1.2.2
 opencensus-ext-azure==1.1.7
 pydantic==1.9.0
 requests==2.28.1
